introduce RenderTemplateService, call RiskIssuesModel not from constructor, adding spock dependencies
