+++ pom.xml
@@ -139,6 +139,21 @@
             <version>1.7</version>
             <scope>provided</scope>
         </dependency>
+        <dependency>
+            <groupId>org.spockframework</groupId>
+            <artifactId>spock-core</artifactId>
+            <version>0.7-groovy-2.0</version>
+        </dependency>
+        <dependency>
+            <groupId>org.codehaus.groovy</groupId>
+            <artifactId>groovy</artifactId>
+            <version>2.0.1</version>
+        </dependency>
+        <dependency>
+            <groupId>cglib</groupId>
+            <artifactId>cglib-nodep</artifactId>
+            <version>2.1_3</version>
+        </dependency>
     </dependencies>
     <build>
         <plugins>
@@ -181,7 +196,7 @@
 
 	<reporting>
 		<plugins>
-			<plugin>
+<!--			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-pmd-plugin</artifactId>
 				<version>3.4</version>
@@ -190,7 +205,7 @@
 				<groupId>org.codehaus.mojo</groupId>
 				<artifactId>findbugs-maven-plugin</artifactId>
 				<version>3.0.1</version>
-			</plugin>
+			</plugin>-->
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-checkstyle-plugin</artifactId>
@@ -198,7 +213,29 @@
 			</plugin>
 		</plugins>
 	</reporting>
-    
+
+    <repositories>
+
+        <repository>
+            <id>atlassian-public</id>
+            <url>http://maven.atlassian.com/content/groups/public</url>
+        </repository>
+
+        <repository>
+            <id>repository.springframework.maven.release</id>
+            <name>Spring Framework Maven Release Repository</name>
+            <url>http://maven.springframework.org/release</url>
+        </repository>
+
+    </repositories>
+
+    <pluginRepositories>
+        <pluginRepository>
+            <id>atlassian-public</id>
+            <url>http://maven.atlassian.com/content/groups/public</url>
+        </pluginRepository>
+    </pluginRepositories>
+
     <properties>
         <jdkLevel>1.6</jdkLevel>
         <jira.version>6.1.7</jira.version>
